+++ air/src/passes/translate_from_mir.rs
@@ -592,8 +592,8 @@ fn build_bus_boundary(
             MirValue::PublicInputTable(public_input_table) => Ok(
                 crate::ir::BusBoundary::PublicInputTable(crate::ir::PublicInputTableAccess::new(
                     public_input_table.table_name,
-                    public_input_table.bus_name(),
                     public_input_table.num_cols,
+                    public_input_table.bus_type(),
                 )),
             ),
             // This represents an empty bus