+++ codegen/ace/src/tests/quotient.rs
@@ -66,6 +66,9 @@ pub fn eval_quotient(air: &Air, ace_vars: &AceVars, log_trace_len: u32) -> QuadF
                     let idx = public[&access.name];
                     ace_vars.public[idx][access.index]
                 },
+                Value::PublicInputTable(_) => {
+                    todo!("public input tables are not supported yet (see #399)")
+                },
                 Value::RandomValue(idx) => ace_vars.rand[idx],
             },
             Operation::Add(l, r) => evals[usize::from(l)] + evals[usize::from(r)],